requirements.txt
streamlit>=1.32.0
ultralytics==8.3.40
opencv-python-headless>=4.9.0
pillow>=10.0.0
numpy>=1.24.0,<2.0.0
torch>=2.2.0
torchvision>=0.17.0
plotly>=5.18.0
pandas>=2.0.0